feat(google_maps_flutter): Deprecates renderer selection API
The legacy Google Maps renderer has been turned down, so requesting the legacy renderer is now a no-op. This change deprecates the current API, and replaces it with one that doesn't take a renderer argument and just allows pre-loading parts of the SDK (by calling the native API with the platform default argument).